Move both demos to a their own package each
