+++ EditTextChips/src/main/java/com/novoda/demo/edittextchips/splitwise/Tag.java
@@ -1,4 +1,4 @@
-package com.novoda.demo.edittextchips;
+package com.novoda.demo.edittextchips.splitwise;
 
 import java.io.Serializable;
