+++ EditTextChips/src/main/res/layout/activity_splitwise.xml
@@ -5,7 +5,7 @@
     android:gravity="bottom"
     android:orientation="vertical">
 
-    <com.novoda.demo.edittextchips.SplitwiseTagsView
+    <com.novoda.demo.edittextchips.splitwise.SplitwiseTagsView
         android:id="@+id/tags"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
