+++ EditTextChips/src/main/res/layout/item_view_tag.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.novoda.demo.edittextchips.TagView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.novoda.demo.edittextchips.splitwise.TagView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/name"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
